requirements.txt
streamlit
pandas
plotly
google-auth